requirements.txt
Django==2.2.3
sentry-sdk==0.10.2
djangorestframework==3.10.0
django-cors-headers==3.0.2
